+++ src/components/organisms/common/container/UserListContainer.tsx
@@ -2,17 +2,32 @@ import { ScrollContainer } from "src/components/atoms/common/container/ScrollCon
 import { Member } from "src/components/organisms/common/Member";
 import type { Member as TMember } from "src/types/member";
 
-type Props = { userList: TMember[] };
+type Props = {
+  title: string;
+  userList: TMember[];
+};
 
 export const UserListContainer = (props: Props) => {
-  const { userList } = props;
+  const { title, userList } = props;
   return (
-    <ScrollContainer height={"460px"}>
-      <ul className="grid grid-cols-7 gap-x-5 gap-y-5">
-        {userList.map((user) => {
-          return <Member key={user.id} name={user.name} id={user.id} isCancel={user.isCancel}></Member>;
-        })}
-      </ul>
-    </ScrollContainer>
+    <>
+      <div className="mb-20">
+        <div className="flex flex-row justify-between">
+          <h1 className="text-5xl">{title}</h1>
+          <div className="flex flex-row items-end mb-8">
+            <h1 className="text-lg mr-3">TOTAL</h1>
+            <h1 className="text-5xl">{userList.length}</h1>
+          </div>
+        </div>
+
+        <ScrollContainer height={"460px"}>
+          <ul className="grid grid-cols-7 gap-x-5 gap-y-5">
+            {userList.map((user) => {
+              return <Member key={user.id} name={user.name} id={user.id} isCancel={user.isCancel}></Member>;
+            })}
+          </ul>
+        </ScrollContainer>
+      </div>
+    </>
   );
 };
