+++ src/pages/index.tsx
@@ -23,7 +23,7 @@ const Home = () => {
         <SearchInput value={searchText} placeholder="Search for apps" onChange={onChangeSearchText} />
       </div>
       <div className="pt-20 flex justify-center space-x-10 > *">
-        <DashboardMenu src="/icon_meet.png" title="MEET" onClick={() => onClickMenu("/test")} />
+        <DashboardMenu src="/icon_meet.png" title="MEET" onClick={() => onClickMenu("/monthSummary")} />
         <DashboardMenu src="/icon_shuffle.png" title="SHUFFLE" onClick={() => alert()} />
         <DashboardMenu src="/icon_team.png" title="TEAM" onClick={() => alert()} />
         <DashboardMenu src="/icon_list.png" title="コレスル" onClick={() => alert()} />
